Creacion de comentarios realiza. A falta de probar por fallos en el modelo

+++ src/main/java/es/uniovi/asw/business/CommentaryService.java
@@ -12,7 +12,7 @@ public interface CommentaryService {
 
 	public Commentary findByCitizen(String dni);
 
-	public void save(int idCitizen, int idProposal, String message);
+	public void save(Long idCitizen, Long idProposal, String message);
 
 	public List<Commentary> findByProposal(long id);
 }
